Cleanup script 3

+++ autoibamr.sh
@@ -113,7 +113,7 @@ PREFIX_PATH=${PREFIX/#~\//$HOME\/}
 unset PREFIX
 
 RE='^[0-9]+$'
-if [[ ! "${JOBS}" =~ ${RE} || ${JOBS}<1 ]] ; then
+if [[ ! "${JOBS}" =~ ${RE} || ${JOBS} -lt 1 ]] ; then
   echo "ERROR: invalid number of build processes '${JOBS}'"
   exit 1
 fi
@@ -147,7 +147,6 @@ BAD="\033[1;31m"
 GOOD="\033[1;32m"
 WARN="\033[1;35m"
 INFO="\033[1;34m"
-BOLD="\033[1m"
 
 ################################################################################
 # Define autoibamr helper functions
@@ -160,7 +159,7 @@ prettify_dir() {
 cecho() {
     # Display messages in a specified color
     COL=$1; shift
-    echo -e "${COL}$@\033[0m"
+    echo -e "${COL}$*\033[0m"
 }
 
 default () {
@@ -277,11 +276,11 @@ download_archive () {
     ARCHIVE_FILE=$1
 
     # Prepend MIRROR to SOURCE (to prefer) mirror source download
-    if [ ! -z "${MIRROR}" ]; then
+    if [ -n "${MIRROR}" ]; then
         SOURCE="${MIRROR} ${SOURCE}"
     fi
 
-    for DOWNLOADER in ${DOWNLOADERS[@]}; do
+    for DOWNLOADER in ${DOWNLOADERS}; do
     for source in ${SOURCE}; do
         # verify_archive:
         # * Skip loop if the ARCHIVE_FILE is already downloaded
@@ -389,16 +388,6 @@ package_fetch () {
         else
             cd ${NAME}
             svn up
-            cd ..
-        fi
-    elif [ ${PACKING} = "bzr" ]; then
-        cd ${UNPACK_PATH}
-        # Suitably branch or update bzr repositories
-        if [ ! -d ${NAME} ]; then
-            bzr branch ${SOURCE}${NAME}
-        else
-            cd ${NAME}
-            bzr pull
             cd ..
         fi
     fi
@@ -602,7 +591,7 @@ guess_platform() {
             esac
         fi
 
-    elif [ ! -z "${CRAYOS_VERSION}" ]; then
+    elif [ -n "${CRAYOS_VERSION}" ]; then
         echo cray
 
     elif [ -f /etc/os-release ]; then
@@ -908,7 +897,7 @@ cecho ${INFO} "Compiler Variables:"
 echo
 
 # CC test
-if [ ! -n "${CC}" ]; then
+if [ -z "${CC}" ]; then
     if builtin command -v mpicc > /dev/null; then
         cecho ${WARN} "CC  variable not set, but default mpicc  found."
         export CC=mpicc
@@ -922,7 +911,7 @@ else
 fi
 
 # CXX test
-if [ ! -n "${CXX}" ]; then
+if [ -z "${CXX}" ]; then
     if builtin command -v mpicxx > /dev/null; then
         cecho ${WARN} "CXX variable not set, but default mpicxx found."
         export CXX=mpicxx
@@ -936,7 +925,7 @@ else
 fi
 
 # FC test
-if [ ! -n "${FC}" ]; then
+if [ -z "${FC}" ]; then
     if builtin command -v mpif90 > /dev/null; then
         cecho ${WARN} "FC  variable not set, but default mpif90 found."
         export FC=mpif90
@@ -950,7 +939,7 @@ else
 fi
 
 # FF test
-if [ ! -n "${FF}" ]; then
+if [ -z "${FF}" ]; then
     if builtin command -v mpif77 > /dev/null; then
         cecho ${WARN} "FF  variable not set, but default mpif77 found."
         export FF=mpif77
@@ -1150,13 +1139,13 @@ for PACKAGE in ${PACKAGES[@]}; do
         # Clean src after install
         if [ ${INSTANT_CLEAN_SRC_AFTER_INSTALL} = ON ]; then
             if [ -f ${DOWNLOAD_PATH}/${NAME}${PACKING} ]; then
-                rm -f ${DOWNLOAD_PATH}/${NAME}${PACKING}
+                rm -f ${DOWNLOAD_PATH:?}/${NAME}${PACKING}
             fi
         fi
 
         # Clean unpack directory after install
         if [ ${INSTANT_CLEAN_UNPACK_AFTER_INSTALL} = ON ]; then
-            rm -rf ${UNPACK_PATH}/${EXTRACTSTO}
+            rm -rf ${UNPACK_PATH:?}/${EXTRACTSTO}
         fi
     else
         if [ ! -z "${LOAD}" ]; then
